📂 Files relocation to avoid having too many dirs
