--- src/__tests__/index.test.tsx
@@ -1,3 +1,0 @@
-it("Sample test", () => {
-  expect(true).toBeTruthy();
-});
