--- src/app/[lang]/components/MyButton.tsx
@@ -1,5 +1,0 @@
-export default function MyButton() {
-  return (
-    <button className="bg--primary p--1 round--big"> Simple button </button>
-  );
-}
