--- src/app/[lang]/hello.mdx
@@ -1,9 +1,0 @@
-import MyButton from './components/MyButton.tsx';
-
-# Hello, Next.js, Iam MDX componente!
-
-I am also can be directly a pahe
-
-* route_name/page.mdx
-<MyButton/>
-###  Finish of MDX componente!
