+++ src/app/[lang]/layout.tsx
@@ -1,6 +1,6 @@
 import { i18n } from "../../../i18n-config";
-import TheHeader from "./components/TheHeader";
-import TheFooter from "./components/TheFooter";
+import TheHeader from "@/app/components/TheHeader";
+import TheFooter from "@/app/components/TheFooter";
 
 export async function generateStaticParams() {
   return i18n.locales.map((locale) => ({ lang: locale }));
