+++ src/app/components/TheFooter/styles.module.scss
