+++ src/app/components/TheNav/TheNav.stories.tsx
@@ -1,0 +1,9 @@
+import { ComponentMeta, ComponentStory } from "@storybook/react";
+import TheNav from "@/app/components/TheNav";
+
+export default {
+  title: "Components/TheNav",
+  component: TheNav,
+} as ComponentMeta<typeof TheNav>;
+
+export const Default: ComponentStory<typeof TheNav> = () => <TheNav />;
