+++ src/app/components/ui/SearchButton/SearhButton.stories.tsx
@@ -1,5 +1,5 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
-import SearchButton from "@/app/[lang]/components/ui/SearchButton";
+import SearchButton from "@/app/components/ui/SearchButton";
 import { action } from "@storybook/addon-actions";
 
 export default {
