+++ src/app/components/ui/SectionTitle/styles.module.scss
