+++ src/styles/_globals.scss
